fix #58 #68
#58 #68
