+++ rpg2d2/Assets/Resources/Scripts/GameManager.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 56d5a2e722320f946b5347fa0d693507
+timeCreated: 1508560856
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
