+++ rpg2d2/Assets/Scene/title.unity
@@ -382,6 +382,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: db304becbb5ea2646a65e8ab32382a4e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  firstSceneName: main
 --- !u!1 &853486354
 GameObject:
   m_ObjectHideFlags: 0
@@ -621,6 +622,46 @@ CanvasRenderer:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1493996533}
+--- !u!1 &1560720401
+GameObject:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 5
+  m_Component:
+  - component: {fileID: 1560720403}
+  - component: {fileID: 1560720402}
+  m_Layer: 0
+  m_Name: GameManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1560720402
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1560720401}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 56d5a2e722320f946b5347fa0d693507, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!4 &1560720403
+Transform:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1560720401}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 42.679813, y: 45.8732, z: 84.29706}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 4
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1796780121
 Prefab:
   m_ObjectHideFlags: 0
@@ -1015,6 +1056,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: db304becbb5ea2646a65e8ab32382a4e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  firstSceneName: main
 --- !u!1 &2110910909
 GameObject:
   m_ObjectHideFlags: 0
